Create ForkedTimeTable.py
This project is called Country Time Table. It is a Python code that creates and sorts a table of countries and their capitals, and also shows the current time for each country. It uses the datetime and pytz modules to manipulate date and time objects and perform time zone conversions. It allows the user to choose the sorting option, which can be either alphabetical or continent. It prints the table in a tabular format, using the pytz.timezone and astimezone functions to convert the UTC time to the local time of each country, and the strftime function to format the datetime object as a string.

Originally authored by PythonDevMaster on replit, I added: 
(Added option to sort by time)
(Added additional USA timezones)